Move getClassNameForAssociation from UnitOfWork to DocumentManager
