+++ lib/Doctrine/ODM/MongoDB/UnitOfWork.php
@@ -2728,48 +2728,6 @@ final class UnitOfWork implements PropertyChangedListener
     }
 
     /**
-     * Gets the class name for an association (embed or reference) with respect
-     * to any discriminator value.
-     *
-     * @internal
-     *
-     * @param FieldMapping              $mapping
-     * @param array<string, mixed>|null $data
-     *
-     * @psalm-return class-string
-     */
-    public function getClassNameForAssociation(array $mapping, $data): string
-    {
-        $discriminatorField = $mapping['discriminatorField'] ?? null;
-
-        $discriminatorValue = null;
-        if (isset($discriminatorField, $data[$discriminatorField])) {
-            $discriminatorValue = $data[$discriminatorField];
-        } elseif (isset($mapping['defaultDiscriminatorValue'])) {
-            $discriminatorValue = $mapping['defaultDiscriminatorValue'];
-        }
-
-        if ($discriminatorValue !== null) {
-            return $mapping['discriminatorMap'][$discriminatorValue]
-                ?? (string) $discriminatorValue;
-        }
-
-        $class = $this->dm->getClassMetadata($mapping['targetDocument']);
-
-        if (isset($class->discriminatorField, $data[$class->discriminatorField])) {
-            $discriminatorValue = $data[$class->discriminatorField];
-        } elseif ($class->defaultDiscriminatorValue !== null) {
-            $discriminatorValue = $class->defaultDiscriminatorValue;
-        }
-
-        if ($discriminatorValue !== null) {
-            return $class->discriminatorMap[$discriminatorValue] ?? $discriminatorValue;
-        }
-
-        return $mapping['targetDocument'];
-    }
-
-    /**
      * Creates a document. Used for reconstitution of documents during hydration.
      *
      * @psalm-param class-string<T> $className
